Migrated ConcurrentLinkedQueue receiver.
